schwab-api/pyproject.toml
[project]
name = "schwab-api"
version = "0.1.0"
description = "Clean Python client for Schwab trading API with focus on options orders"
requires-python = ">=3.10"
dependencies = [
    "requests>=2.31.0",
    "python-dotenv>=1.0.0",
    "flask>=3.0.0",
    "pyopenssl>=23.0.0",
]

[project.optional-dependencies]
dev = ["pytest>=7.4.0", "black>=23.0.0", "ruff>=0.1.0"]

[tool.black]
line-length = 100
target-version = ["py310"]

[tool.ruff]
line-length = 100
select = ["E", "F", "I"]
